+++ common/src/main/resources/bandagez.mixins.json
@@ -6,7 +6,6 @@
   "compatibilityLevel": "JAVA_18",
   "mixins": [],
   "client": [
-    "MixinMinecraft"
   ],
   "server": [],
   "injectors": {
