+++ fabric/src/main/java/ru/ancndz/bandagez/loot/GrassDropModifier.java
@@ -1,4 +1,4 @@
-package ru.ancndz.bandagez;
+package ru.ancndz.bandagez.loot;
 
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootTableProvider;
