+++ fabric/src/main/resources/bandagez.fabric.mixins.json
@@ -6,7 +6,6 @@
   "compatibilityLevel": "JAVA_21",
   "mixins": [],
   "client": [
-    "MixinTitleScreen"
   ],
   "server": [],
   "injectors": {
