--- forge/src/main/java/ru/ancndz/bandagez/BandagezModConfig.java
@@ -1,35 +1,0 @@
-package ru.ancndz.bandagez;
-
-import net.minecraftforge.common.ForgeConfigSpec;
-import net.minecraftforge.fml.config.ModConfig;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import org.apache.commons.lang3.tuple.Pair;
-
-public class BandagezModConfig {
-
-    public static void init(FMLJavaModLoadingContext context) {
-        context.registerConfig(ModConfig.Type.CLIENT, BandagezModConfig.clientSpec);
-    }
-
-    public static final class Client {
-
-        public final ForgeConfigSpec.BooleanValue showParticles;
-
-        Client(final ForgeConfigSpec.Builder builder) {
-            showParticles = builder
-                    .comment("Show particles on bleeding effects")
-                    .translation("bandagez.config.show_bleeding_particles")
-                    .define("showParticles", true);
-        }
-    }
-
-    public static final ForgeConfigSpec clientSpec;
-
-    public static final Client CLIENT;
-
-    static {
-        final Pair<Client, ForgeConfigSpec> specPair = new ForgeConfigSpec.Builder().configure(BandagezModConfig.Client::new);
-        clientSpec = specPair.getRight();
-        CLIENT = specPair.getLeft();
-    }
-}
