+++ neoforge/src/main/java/ru/ancndz/bandagez/loot/GrassDropModifier.java
@@ -1,4 +1,4 @@
-package ru.ancndz.bandagez;
+package ru.ancndz.bandagez.loot;
 
 import com.google.common.base.Suppliers;
 import com.mojang.serialization.MapCodec;
@@ -6,14 +6,12 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.LootContext;
-import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
-import net.minecraftforge.common.loot.IGlobalLootModifier;
-import net.minecraftforge.common.loot.LootModifier;
-import net.minecraftforge.registries.DeferredRegister;
-import net.minecraftforge.registries.ForgeRegistries;
-import org.jetbrains.annotations.NotNull;
-import ru.ancndz.bandagez.loot.CommonGrassDropLogic;
+import net.neoforged.neoforge.common.loot.IGlobalLootModifier;
+import net.neoforged.neoforge.common.loot.LootModifier;
+import net.neoforged.neoforge.registries.DeferredRegister;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
+import ru.ancndz.bandagez.BandagezMod;
 
 import java.util.function.Supplier;
 
@@ -22,7 +20,7 @@ public class GrassDropModifier extends LootModifier {
     public static final String GRASS_DROP_MODIFIER = "grass_drop_modifier";
 
     public static final DeferredRegister<MapCodec<? extends IGlobalLootModifier>> LOOT_MODIFIERS =
-            DeferredRegister.create(ForgeRegistries.Keys.GLOBAL_LOOT_MODIFIER_SERIALIZERS, BandagezMod.MODID);
+            DeferredRegister.create(NeoForgeRegistries.Keys.GLOBAL_LOOT_MODIFIER_SERIALIZERS, BandagezMod.MODID);
 
     public static final Supplier<MapCodec<GrassDropModifier>> CODEC_SUPPLIER = Suppliers
             .memoize(() -> RecordCodecBuilder.mapCodec(inst -> codecStart(inst).apply(inst, GrassDropModifier::new)));
@@ -32,9 +30,8 @@ public class GrassDropModifier extends LootModifier {
     }
 
     @Override
-    protected @NotNull ObjectArrayList<ItemStack>
-            doApply(LootTable table, ObjectArrayList<ItemStack> generatedLoot, LootContext context) {
-        return CommonGrassDropLogic.apply(table, generatedLoot, context);
+    protected ObjectArrayList<ItemStack> doApply(ObjectArrayList<ItemStack> generatedLoot, LootContext lootContext) {
+        return CommonGrassDropLogic.apply(generatedLoot, lootContext);
     }
 
     @Override
